PR feedback, move code to module, restore unintended changes
