+++ lib/capybara/node/element.rb
@@ -56,12 +56,9 @@ module Capybara
       # @return [String]              The text of the element
       #
       def text(type = nil, normalize_ws: false)
-        unless session_options.ignore_hidden_elements || session_options.visible_text_only
-          type ||= :all
-        end
-
-        extracted_text = synchronize { type == :all ? base.all_text : base.visible_text }
-        normalize_ws ? extracted_text.gsub(/[[:space:]]+/, ' ').strip : extracted_text
+        type ||= :all unless session_options.ignore_hidden_elements || session_options.visible_text_only
+        txt = synchronize { type == :all ? base.all_text : base.visible_text }
+        normalize_ws ? txt.gsub(/[[:space:]]+/, ' ').strip : txt
       end
 
       ##
